+++ server/src/prefect_server/services/zombie_killer/__init__.py
@@ -1,0 +1,4 @@
+# Licensed under the Prefect Community License, available at
+# https://www.prefect.io/legal/prefect-community-license
+
+from prefect_server.services.zombie_killer.zombie_killer import ZombieKiller