+++ server/Changelog.md
@@ -9,6 +9,7 @@
 ### Enhancements
 
 - new cli command--`prefect-server dev build` will build prefect server images from source locally - [#2227](https://github.com/PrefectHQ/prefect/pull/2227)
+- Zombie Killer extension brought to server
 
 ### Infrastructure